Interface

+++ Biblioteca/src/Biblio.Core/Autor.cs
@@ -2,7 +2,7 @@ namespace Biblio.Core;
 
 public class Autor
 {
-    public required uint IdAutor { get; set; }
+    public required ushort IdAutor { get; set; }
     public required string Nombre { get; set; }
     public required string Apellido { get; set; }
 }
